ReadMe
In this notebook, I use a fictional dataset to demonstrate a very common data manipulation problem I run into in Audit Analytics projects: given the weight factors across many groups, how to calculate the percentage weight for each members within a group. For example, in the following dataset, the column "Contribution_Factor" is used to allocate the department total Bonus: the total bonus for a department will be allocated to the employees in this department proportional to the employee's contribution factor.